Switch to `cabal new-test` for nix-style script generator
We can finally drop the `new-test` emulation since now there's actually
the real thing in cabal.

This has the benefit of providing an augmented environment in which to
run tests which is important for tests to find their package
data-files (e.g. via `Paths_` modules - and hence fixes #75) as well as
tests being able to locate internal executables declared via
`build-tools` via `$PATH`