Document copy-fields, and make local-ghc-options always fire
I attempted to use the `copy-fields` feature the other day only to
realize that I have no idea how it is intended to work. This is my
attempt at adding documentation for it.

Along the way, I noticed that local-ghc-options only ever triggered
if you specify `copy-fields: some`, which seems off. I factored
out the code that generates `local-ghc-options` so that it fires
irrespective of the `copy-fields` setting.