Rename prjOrigFields to prjOtherFields, change parseProject's semantics
This changes the semantics of `Project`'s `prjOtherFields`
(previously `prjOrigFields`) to reflect its intended purpose of
being a catch-all field for everything that is not already covered by
an existing field of `Project`. I've tweaked golden test in
`cabal-parsers-golden` to ensure that these semantics are being
upheld by `parseProject`.

Fixes #367.