GitHub Actions: Allowed failure cleanup
Unfortunately, omitting a property like `continue-on-error` from all
`include` entry causes only one job to be run, as observed in
https://github.com/haskell-CI/haskell-ci/pull/425#issuecomment-747047769.
To work around this, we now enumerate `continue-on-error` in each
`include` entry, verbose as it may be.

While I was in town, I renamed the property in each `include` entry
from `continue-on-error` to `allow-failure`, as this makes it less likely
to confuse with the top-level `continue-on-error` property (a separate
thing entirely).