GitHub Actions: include list of GHC versions in matrix
Alongside "include" include the list of GHC versions in the matrix.
That is, instead of:

  matrix:
    include:
      - ghc: 8.10.3
        allow-failures: false
      - ghc: 8.8.4
        allow-failures: false

haskell-ci now produces:

  matrix:
    ghc:
      - 8.10.3
      - 8.8.4
    include:
      - ghc: 8.10.3
        allow-failures: false
      - ghc: 8.8.4
        allow-failures: false

This approach makes it possible to add another "dimension" to the
matrix using only `--github-patches`, by adding another list of
values.  As a concrete example, the *notmuch* library must be tested
against various versions of libnotmuch:

  matrix:
    notmuch: ["0.29.3", "0.30", "0.31.3", "git"]
    ghc:
      - 8.10.3
      - 8.8.4
    include:
      - ghc: 8.10.3
        allow-failures: false
      - ghc: 8.8.4
        allow-failures: false

The resulting matrix is the cartesian product of the values from the
"ghc" and other list(s).  The additional matrix line(s), and
corresponding build steps, can be introduced via --github-patch.

The "include" maps propagate the "allow-failures" field to the
matrix legs with matching "ghc" version (evidence:
https://github.com/purebred-mua/hs-notmuch/actions/runs/466042178).