Avoid Windows temporary file locking in patchYAML
On Windows, it seems that a temporary file is deemed to be in use between
an `hFlush` and an `hClose`. This causes problems for the `patchYAML` function,
which tries to `patch` a file in between `hFlush` and `hClose`.

The fix is simple enough: just call `patch` after `hClose`, which avoids the
problem entirely. This works on Windows and Unix-like OSes alike.